+++ src/command/check/check.ts
@@ -13,13 +13,11 @@ import { render } from "../render/render-shared.ts";
 import { JupyterCapabilities } from "../../core/jupyter/types.ts";
 import { jupyterCapabilities } from "../../core/jupyter/capabilities.ts";
 import {
-  JupyterCapabilities,
-  jupyterCapabilities,
   jupyterCapabilitiesMessage,
   jupyterInstallationMessage,
   jupyterUnactivatedEnvMessage,
   pythonInstallationMessage,
-} from "../../core/jupyter/capabilities.ts";
+} from "../../core/jupyter/jupyter-shared.ts";
 import { completeMessage, withSpinner } from "../../core/console.ts";
 import {
   KnitrCapabilities,
