+++ src/command/render/render-shared.ts
@@ -1,3 +1,9 @@
+/*
+* render-shared.ts
+*
+* Copyright (C) 2020 by RStudio, PBC
+*
+*/
 import {
   projectContext,
   projectContextForDirectory,
