+++ src/core/jupyter/jupyter.ts
@@ -110,10 +110,10 @@ import {
 } from "../../config/constants.ts";
 import {
   isJupyterKernelspec,
-  JupyterKernelspec,
   jupyterKernelspec,
   jupyterKernelspecs,
 } from "./kernels.ts";
+import { JupyterKernelspec } from "./types.ts";
 import { figuresDir, inputFilesDir } from "../render.ts";
 import { lines } from "../text.ts";
 import { readYamlFromMarkdown, readYamlFromMarkdownFile } from "../yaml.ts";
