+++ src/core/jupyter/types.ts
@@ -1,0 +1,31 @@
+/*
+* types.ts
+*
+* Copyright (C) 2020 by RStudio, PBC
+*
+*/
+
+// deno-lint-ignore-file camelcase
+
+export interface JupyterCapabilities {
+  versionMajor: number;
+  versionMinor: number;
+  versionPatch: number;
+  versionStr: string;
+  execPrefix: string;
+  executable: string;
+  conda: boolean;
+  pyLauncher: boolean;
+  // deno-lint-ignore camelcase
+  jupyter_core: string | null;
+  nbformat: string | null;
+  nbclient: string | null;
+  ipykernel: string | null;
+  kernels?: JupyterKernelspec[];
+}
+
+export interface JupyterKernelspec {
+  name: string;
+  language: string;
+  display_name: string;
+}
